Fix sign-conversion errors

+++ Projects/FileEncryption/src/core/aes_gcm_enc.cpp
@@ -192,7 +192,7 @@ Result AesGcm::EncryptBatch() {
 
     /* Update progress */
 
-    progress_cur_ += kBuffSize * kBlockSize;
+    progress_cur_ += static_cast<int64_t>(kBuffSize * kBlockSize);
 
     ReportProgress();
